Move Settings dialog to another folder
